First Commit. The project is migrated from MATLAB code for a better running speed. (Shouldn't I use C/C++ if speed is what I concerns?) Use pypy for even better performance. Please pay attention that this version has not been tested.
Most functions are stored in "simulation.py", while in BPNN module gives an class PIDwithBPNN and Event module gives an class Event. Event handler is now a method of class Event.
Frame module is currently deprecated and my be deleted later.
GraphPlot is currently empty but will be implemented in the future. So far matplotlib is not available for pypy so use this module carefully.